### 2021-10-04 8d4d59d
Update missing DOIs
