+++ paper.bib
@@ -47,6 +47,7 @@
 @ARTICLE{pymoo,
     author={J. {Blank} and K. {Deb}},
     journal={IEEE Access},
+    doi = {10.1109/ACCESS.2020.2990567},
     title={Pymoo: Multi-Objective Optimization in Python},
     year={2020},
     volume={8},
@@ -61,7 +62,8 @@
   volume={37},
   pages={123--141},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.rser.2014.05.007}
 }
 
 @inproceedings{Mehrfeld.HPC.2020,
@@ -83,6 +85,7 @@
 	author = {Storek, Thomas and Esmailzadeh, Asad and Mehrfeld, Philipp and Schumacher, Markus and Baranski, Marc and Müller, Dirk},
 	year = {2019},
 	pages = {8},
+	doi = {10.26868/25222708.2019.210992}
 }
 
 @inproceedings{vering_borges,
@@ -111,7 +114,7 @@
   journal={submitted to: Energy Conversion and Management},
   year={2021},
   publisher={Elsevier},
-  note = {Submitted to}
+  doi = {10.3384/ecp21181561}
 }
 
 @misc{dymola_software,
