+++ paper.bib
@@ -105,16 +105,16 @@
     booktitle = {14th {International} {Modelica} {Conference}},
     year = {2021},
     month = september,
-    note = {Accepted for publication}
+    doi = {10.3384/ecp21181561}
 }
 
 @article{vering_wullhorst_ecm,
   title={Towards an Integrated Design of Heat Pump Systems: Application of Process Intensification using Two-Stage Optimization},
   author={Vering, Christian and Wüllhorst, Fabian and Mehrfeld, Philipp and Müller, Dirk},
-  journal={submitted to: Energy Conversion and Management},
+  journal={Energy Conversion and Management},
   year={2021},
   publisher={Elsevier},
-  doi = {10.3384/ecp21181561}
+  note = {Submitted to}
 }
 
 @misc{dymola_software,
