### 2021-10-04 4721e82
Add case for when optimization fails before first iteration
