### 2021-10-05 #3800 @kyleniemeyer
Hi @FWuellhorst, I noticed a few remaining (minor) issues with the paper that you could resolve while we are getting started:

- In the second paragraph, you have references with nested parentheses :"(e.g. (Blank & Deb, 2020...". You can avoid this by changing the citation to `(e.g., @2020SciPy-NMeth; @dlib09; @pymoo)`.
- In the third paragraph: "it's usage" -> "its usage"
- In the first paragraph of Statement of Need, you have a period after "digital twins" but prior to the references, which are part of that sentence.
- The two Vering et al. (2021) references seem to be missing some bibliographic information.
