### 2021-10-06 aef69d3
Update based on joss review
