### 2021-10-06 045cc1f
Update missing bibliographic information
