+++ paper.bib
@@ -89,13 +89,13 @@
 }
 
 @inproceedings{vering_borges,
-	address = {Bruges, Belgium},
+    	address = {Bruges, Belgium},
 	title = {Digital Twin Design with On-Line Calibration for HVAC Systems in Buildings},
 	language = {en},
 	author = {Vering, Christian and Borges, Sebastian Coakly, Daniel and Krützfeldt, Hannah and Mehrfeld, Philipp and Müller, Dirk},
+	booktitle = {Proceedings of the 17th {IBPSA} {Conference}},
 	year = {2021},
-	pages = {8},
-    note = {Accepted for publication}
+	pages = {8}
 }
 
 @inproceedings{ModelicaConferenceWullhorst,
@@ -111,10 +111,9 @@
 @article{vering_wullhorst_ecm,
   title={Towards an Integrated Design of Heat Pump Systems: Application of Process Intensification using Two-Stage Optimization},
   author={Vering, Christian and Wüllhorst, Fabian and Mehrfeld, Philipp and Müller, Dirk},
-  journal={Energy Conversion and Management},
+  journal={Submitted to: Energy Conversion and Management},
   year={2021},
-  publisher={Elsevier},
-  note = {Submitted to}
+  publisher={Elsevier}
 }
 
 @misc{dymola_software,
