+++ paper.bib
@@ -108,12 +108,13 @@
     doi = {10.3384/ecp21181561}
 }
 
-@article{vering_wullhorst_ecm,
-  title={Towards an Integrated Design of Heat Pump Systems: Application of Process Intensification using Two-Stage Optimization},
+@inproceedings{vering_wullhorst_ecm,
+  title={Towards an Integrated Design of Heat Pump Systems: Application and Assessment of Process Intensification using Two-Stage Optimization},
   author={Vering, Christian and Wüllhorst, Fabian and Mehrfeld, Philipp and Müller, Dirk},
-  journal={Submitted to: Energy Conversion and Management},
+  booktitle={{The 34th International Conference On Efficiency, Cost, Optimization, Simulation and Environmental Impact of Energy System}},
   year={2021},
-  publisher={Elsevier}
+  month = june,
+  language = {en}
 }
 
 @misc{dymola_software,
