### 2021-10-06 #3800 @FWuellhorst
@kyleniemeyer thank you for your review!
I corrected the listed issues.
Two remarks regarding the missing information in the references of Vering et al. (2021).
1. The conference article regarding digital twins is part of the 17th IBPSA Conference, which already took place. The proceedings are not published yet. A DOI is not yet present.
2. We submitted an article to Energy Conversion and Management, which is currently under review. As the article extends our already published conference paper at the ECOS 2021 (https://www.ebc.eonerc.rwth-aachen.de/go/id/dncb/file/826177?lidx=1), I will refer to the conference instead. Again, a DOI is not yet present.
