### 2021-10-19 #3800 @fraukewiese
@kramea @jordanperr – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

The submission under review here is [AixCaliBuHA: Automated calibration of building and HVAC systems https://github.com/RWTH-EBC/AixCaliBuHA](https://github.com/RWTH-EBC/AixCaliBuHA)
