### 2021-10-27 e8805ef
Fix setup.py, update version number for new release
