### 2021-11-17 #3861 @ralphevins
@fraukewiese I have a question.
 The software in question is an API to Dymola, a (very expensive) commercial product.
As is, it opens up Dymola in the system and returns 'Can't find AixLib library' error.
The free version of Dymola cannot support AixLib because the trial version supports limited state-space equations and full models in the library can't be translated or simulated.
So unfortunately it looks as if we must reject, as it's not possible to review (or use) the library without a Dymola license?
