### 2021-11-18 #3861 @FWuellhorst
@fraukewiese @ralphevins :
First of all thanks for taking the time to review our software code.

To explain the issue here: You probably started with the first example, which is just there to help users analyze their system and identify inputs, outputs, and parameters of the model they need to calibrate. We think users should think about what they want to calibrate before they actually do it. This would be possible using our main simulation tool, the `FMU_API`, as well (https://fmi-standard.org/). I apologize that we made this first example for Dymola users only. We will update it directly.
Regarding our Simulation-APIs, note from the paper:

> Currently, the package supports the calibration of Functional Mock-up Units (FMUs) based on the Functional Mock-up Interface (FMI) standard (Modelica Association Project, 2021) as well as Modelica models through the python_interfaceof the Software Dymola (Dassault Systems, 2021).

So, the review of code and software should be possible using the `FMU_API`. If the presence of the `DymolaAPI` hinders the publication process, we can remove this API from the source code.

All other examples and most importantly the calibration work with the open FMI standard using the interface to `fmpy`.
We provide two example models. One can be executed on both Linux and windows (Example B) and one only on windows (Example A). If you have trouble running these examples please let me know. I will also update the examples README.md to clarify the points I just made.

Sorry again for any inconvenience due to the bad example description. We hope you will still consider the software for publication after this clarification and the update of the example documentation.
