### 2021-11-18 RWTH-EBC/AixCaliBuHA#23 @FWuellhorst
Clarify example description
As stated in this [issue](https://github.com/openjournals/joss-reviews/issues/3861), the first example is not directly useable if no Dymola license is present. Hence, I will create an example solely based on the `FMU_API`.

- [x] Clarify which examples can be run on which machine.
- [x] Create a new example `e1` for the `PumpAndValve` fmu model
