### 2021-11-18 d5470b6
Refactor files and add detailed description
