+++ examples/README.md
@@ -5,34 +5,37 @@ This folder contains several example files which help with the understanding of 
 # Getting started
 
 While these examples should run in any IDE, we advise using PyCharm.
-Before being able to run these examples, be sure to:
+To execute these examples, be sure to:
 
 1. Create a clean environment of python 3.7 or 3.8. In Anaconda run: `conda create -n py38_ebcpy python=3.8`
 2. Activate the environment in your terminal. In Anaconda run: `activate py38_ebcpy`
 3. Clone the library using `git clone --recurse-submodules https://github.com/RWTH-EBC/AixCaliBuHA`
-4. Clone the AixLib in order to use the models: `git clone https://github.com/RWTH-EBC/AixLib`
-   Also check if you're on development using `cd AixLib && git status && cd ..`
 5. Install the library using `pip install -e AixCaliBuHA`
-   
+
+We have two models to show the calibration process for different components inside typical building and HVAC systems.
+
+- Example model A: This model is of a heat pump system supplying heat to a room using a radiator. The models are based on the `AixLib`.
+- Example model B: This model is of a pump and a valve from the Modelica Standard Library. 
+
+Currently, example A runs only on windows. Example B runs on both linux and windows.
+To follow a specific example A or B, execute the first and second example for the case, for instance `e1_A`.
+The examples 3-5 are written for both, so just alter the parameter in the `if __name__ == '__main__'` section, for instance `EXAMPLE = A` to `EXAMPLE = B`.
+
 # What can I learn in the examples?
 
-## `e1_energy_system_analysis.py`
+## `e1_A_energy_system_analysis.py` and `e1_B_energy_system_analysis.py`
 
-1. Learn how to analyze your energy system
-2. Improve your `DymolaAPI` knowledge
+1. Learn how to analyze the model of your energy system
+2. Improve your `SimulationAPI` knowledge
 3. Improve your skill-set on `TimeSeriesData`
 
-## `e2_1_optimization_problem_definition.py`
+## `e2_A_optimization_problem_definition.py` and `e2_B_optimization_problem_definition.py`
 
 1. Learn how to formulate your calibration problem using our data_types
 2. Get to know `TunerParas`
 3. Get to know `Goals`
 4. Get to know `CalibrationClass`
 5. Learn how to merge multiple classes
-
-## `e2_2_optimization_problem_definition.py`
-
-1. Deepen your understanding of data_types in AixCaliBuHA
 
 ## `e3_sensitivity_analysis_example.py`
 
