### 2021-11-18 391ba7f
Set default to example B
