### 2021-11-18 1cd0eb5
Add data and refactor second example
