+++ examples/__init__.py
@@ -4,7 +4,7 @@ Module containing examples on how to use AixCaliBuHA
 import pathlib
 import sys
 from ebcpy import FMU_API
-from examples import e2_A_optimization_problem_definition, e2_B_another_data_types_example
+from examples import e2_A_optimization_problem_definition, e2_B_optimization_problem_definition
 
 
 def setup_fmu(example="B"):
@@ -31,4 +31,4 @@ def setup_calibration_classes(example="B", multiple_classes=True):
     """Setup the CalibrationClasses used in all examples and tests."""
     if example == "A":
         return e2_A_optimization_problem_definition.main(multiple_classes=multiple_classes)
-    return e2_B_another_data_types_example.setup_calibration_classes()
+    return e2_B_optimization_problem_definition.setup_calibration_classes()
