### 2021-11-18 730fbf3
Adjust examples
