+++ examples/__init__.py
@@ -31,4 +31,4 @@ def setup_calibration_classes(example="B", multiple_classes=True):
     """Setup the CalibrationClasses used in all examples and tests."""
     if example == "A":
         return e2_A_optimization_problem_definition.main(multiple_classes=multiple_classes)
-    return e2_B_optimization_problem_definition.setup_calibration_classes()
+    return e2_B_optimization_problem_definition.main(multiple_classes=multiple_classes)
