+++ examples/e2_B_optimization_problem_definition.py
@@ -1,87 +1,173 @@
 """
-Example file for the data_types module. The usage of classes inside
-the data_types module should be clear when looking at the examples.
-If not, please raise an issue.
+Goals of this part of the examples:
+1. Learn how to formulate your calibration problem using our data_types
+2. Get to know `TunerParas`
+3. Get to know `Goals`
+4. Get to know `CalibrationClass`
+5. Learn how to merge multiple classes
 """
+# Start by importing all relevant packages
 import pathlib
+# Imports from ebcpy
 from ebcpy import TimeSeriesData
-from aixcalibuha import CalibrationClass, TunerParas, Goals
+# Imports from AixCaliBUhA
+from aixcalibuha import TunerParas, Goals, \
+    CalibrationClass
+from aixcalibuha.data_types import merge_calibration_classes
 
 
-def setup_goals():
+def main(
+        statistical_measure="NRMSE",
+        multiple_classes=True
+):
     """
-    Example setup of the Goals object.
-    First, some simulated and measured target data is loaded from the
-    example data.
-    Then the goals object is instantiated. Please refer to the
-    Goals documentation on the meaning of the parameters.
+    Arguments of this example:
 
-
-    :return: Goals object
-    :rtype: aixcalibuha.Goals
+    :param str statistical_measure:
+        Measure to calculate the scalar of the objective,
+        One of the supported methods in
+        ebcpy.utils.statistics_analyzer.StatisticsAnalyzer
+        e.g. RMSE, MAE, NRMSE
+    :param bool multiple_classes:
+        If False, all CalibrationClasses will have the
+        same name
     """
-    # As the examples should work, and the cal_class example uses the other examples,
-    # we will test it here:
-    data_dir = pathlib.Path(__file__).parent.joinpath("data")
-    sim_target_data = TimeSeriesData(data_dir.joinpath("PumpAndValveSimulation.hdf"),
-                                     key="examples")
-    meas_target_data = TimeSeriesData(data_dir.joinpath("PumpAndValve.hdf"),
-                                      key="examples")
-
-    # Setup three variables for different format of setup
-    variable_names = {"TCap": ["TCapacity", "heatCapacitor.T"],
-                      "TPipe": {"meas": "TPipe", "sim": "pipe.T"}}
-
-    # Convert index to float to match the simulation output
-    meas_target_data.to_float_index()
-    # Setup the goals object
-    goals = Goals(meas_target_data=meas_target_data,
-                  variable_names=variable_names,
-                  statistical_measure="NRMSE",
-                  weightings=[0.7, 0.3])
-    goals.set_sim_target_data(sim_target_data)
-    return goals
-
-
-def setup_calibration_classes():
-    """
-    Example setup of a list calibration classes.
-    The measured data of the setup_goals example can
-    be segmentized into two classes. You can either use
-    classes from the segmentizer package or manually define
-    classes of interest to you. In this example the we have
-    a manual segmentation, as the example is fairly small.
-
-    :return: List of calibration classes
-    :rtype: list
-    """
-    # Specify the tuner parameters and goals
+    # ######################### Tuner Parameter ##########################
     tuner_paras = TunerParas(names=["speedRamp.duration", "valveRamp.duration"],
                              initial_values=[0.1, 0.1],
                              bounds=[(0.1, 10), (0.1, 10)])
 
+    print(tuner_paras)
+    print("Names of parameters", tuner_paras.get_names())
+    print("Initial values", tuner_paras.get_initial_values())
+    # Scaling (will be done internally)
+    print("Scaled initial values:\n", tuner_paras.scale(tuner_paras.get_initial_values()))
+
+    # ######################### Goals ##########################
+    # The evaluation of your goals (or mathematically speaking 'objective function')
[…]
