+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,11 @@
+Thank you for creating a Pull Request for AixCaliBuHA!
+
+Please consider these points before submitting:
+- Mention the issue that this Pull Request intends to solve, and briefly describe what the main changes are.
+- Make sure all new code follows our [Contribution Guidelines]([Documentation](https://ebc.pages.rwth-aachen.de/EBC_all/github_ci/AixCaliBuHA/master/docs/contribution.html))
+- Make sure the code quality does not decrease
+- Make sure the coverage rate does not decrease
+
+If you know someone who could review your code, please use the *Reviewer* function on the right to ask for a review and also assign the PR to them. If not, no worries, we will find someone for you.
+
+Thanks again for your PR! You can delete these first lines before submitting your PR.
