### 2021-12-15 f4e2528
Add semantic release
