+++ docs/source/conf.py
@@ -64,10 +64,11 @@ author = 'EON EBC'
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
-# The short X.Y version.
-version = '0.2'
 # The full version, including alpha/beta/rc tags.
 release = '0.2.2'
+# The short X.Y version.
+version = '.'.join(release.split('.')[:-1])
+
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
