+++ setup.cfg
@@ -1,3 +1,6 @@
 # Inside of setup.cfg
 [metadata]
 description-file = README.md
+
+[semantic_release]
+version_variable = [setup.py:__version__, aixcalibuha/__init__.py:__version__, docs/conf.py:release]
