+++ docs/source/conf.py
@@ -65,7 +65,7 @@ author = 'EON EBC'
 # built documents.
 #
 # The full version, including alpha/beta/rc tags.
-release = '0.2.2'
+release = '0.2.3'
 # The short X.Y version.
 version = '.'.join(release.split('.')[:-1])
 
