+++ setup.cfg
@@ -3,4 +3,4 @@
 description-file = README.md
 
 [semantic_release]
-version_variable = [setup.py:__version__, aixcalibuha/__init__.py:__version__, docs/conf.py:release]
+version_variable = setup.py:__version__,aixcalibuha/__init__.py:__version__,docs/conf.py:release
