### 2021-12-15 168de96
feat: Add semantic release and workflows for github according to #26
