### 2021-12-16 173f131
moved logging from obj to mp_obj trying to fix logging Probelms
