+++ aixcalibuha/calibration/calibrator.py
@@ -14,6 +14,7 @@ from ebcpy.simulationapi import SimulationAPI
 from aixcalibuha.utils import visualizer, MaxIterationsReached
 from aixcalibuha import CalibrationClass, Goals, TunerParas
 import multiprocessing as mp
+import multiprocessing_logging
 
 
 class Calibrator(Optimizer):
@@ -177,7 +178,7 @@ class Calibrator(Optimizer):
                         f"to measurement target data frequency: {mean_freq}")
         self.sim_api.sim_setup.output_interval = mean_freq
 
-    def obj(self, xk, work_id, *args):
+    def obj(self, xk, paras, work_id, *args):
         """
         Default objective function.
         The usual function will be implemented here:
@@ -196,113 +197,234 @@ class Calibrator(Optimizer):
             Objective value based on the used quality measurement
         :rtype: float
         """
-        # Info: This function is called by the optimization framework (scipy, dlib, etc.)
-        #%% Initialize class objects
-        self._current_iterate = xk
-        self._counter += 1
-        # Convert set if multiple goals of different scales are used
-        xk_descaled = self.tuner_paras.descale(xk)
+        use_mp = True
+        if not use_mp:
+            # Info: This function is called by the optimization framework (scipy, dlib, etc.)
+            #%% Initialize class objects
+            self._current_iterate = xk
+            self._counter += 1
+
+            # Convert set if multiple goals of different scales are used
+            xk_descaled = self.tuner_paras.descale(xk)
+
+            # Set initial values of variable and fixed parameters
+            self.sim_api.result_names = self.goals.get_sim_var_names()
+            initial_names = self.tuner_paras.get_names()
+            parameters = self.fixed_parameters.copy()
+            parameters.update({name: value for name, value in zip(initial_names, xk_descaled.values)})
+            # Simulate
+            # pylint: disable=broad-except
+            try:
+                # Generate the folder name for the calibration
+                if self.save_files:
+                    savepath_files = os.path.join(self.sim_api.cd,
+                                                  f"simulation_{self._counter}")
+                    _filepath = self.sim_api.simulate(
+                        parameters=parameters,
+                        return_option="savepath",
+                        savepath=savepath_files,
+                        inputs=self.calibration_class.inputs,
+                        **self.calibration_class.input_kwargs
+                    )
+                    # %% Load results and write to goals object
+                    sim_target_data = data_types.TimeSeriesData(_filepath)
+                else:
+                    sim_target_data = self.sim_api.simulate(
+                        work_id=work_id,
+                        parameters=parameters,
+                        inputs=self.calibration_class.inputs,
+                        **self.calibration_class.input_kwargs
+                    )
+            except Exception as err:
+                if self.fail_on_error:
+                    self.logger.error("Simulation failed. Raising the error.")
+                    raise err
+                self.logger.error(
+                    f"Simulation failed. Returning '{self.ret_val_on_error}' "
+                    f"for the optimization. Error message: {err}"
+                )
+                return self.ret_val_on_error
+
+            self.goals.set_sim_target_data(sim_target_data)
+            # Trim results based on start and end-time of cal class
+            self.goals.set_relevant_time_intervals(self.calibration_class.relevant_intervals)
+
+            #%% Evaluate the current objective
+            # Penalty function (get penalty factor)
+            if self.recalibration_count > 1 and self.apply_penalty:
+                # There is no benchmark in the first iteration or
+                # first iterations were skipped, so no penalty is applied
+                current_tuner_scaled = self.tuner_paras.scale(xk_descaled)
+                penaltyfactor = self.get_penalty(xk_descaled, current_tuner_scaled)
+                # Evaluate with penalty
+                penalty = penaltyfactor
+            else:
+                # Evaluate without penalty
+                penaltyfactor = 1
+                penalty = None
[…]
