### 2022-01-04 781b22b
Don't use semantic release of gitlab
