### 2022-01-05 #3861 @shamsiharis
@FWuellhorst
I am almost done with the review. I was able to follow the tutorial along with the examples. I had two questions though.
1. For the multiple class calibration in the tutorial, running the example in jupyter notebook returns a Simulation Failed Error. The multiple class calibrator runs successfully for the examples via the terminal although returns a maximum iterations reached error.
2. Another error appears at the end of every example, which is related to the FMU_API (FMU_API: Could not terminate fmu instance: fmi2Terminate failed with status 3 (error)).
