### 2022-01-06 RWTH-EBC/AixCaliBuHA#28 @FWuellhorst
Log a warning instead of errors on `MaxIterationsReached`
As pointed out during the [review ](https://github.com/openjournals/joss-reviews/issues/3861) of this codebase, the error log when `MaxIterationsReached` is raised is not appropriate as it's no error. Hence, I will change it to log a warning.
