### 2022-01-06 #3861 @FWuellhorst
@shamsiharis :
Thanks a lot that you take the time to review our code!

Regarding your questions:
1. I will split my answer:
**Simulation error:** Can you raise an issue with the simulation error in AixCaliBuHA? My colleagues and I don't get this error. Please state the error log.
**Maximum iterations reached error:** This was a design choice. Most optimization algorithms determine when to stop. However, setting an explicit maximum number of iterations was a feature request by many users. If an error during calibration occurs, the current best result is saved. Thus, we raise our own error and catch it to exit the optimization gracefully. This could of course be an INFO-log. But we decided on the ERROR log to indicate that the optimization itself did not yet converge. Maybe a warning would be the best log option. I raised an issue: https://github.com/RWTH-EBC/AixCaliBuHA/issues/28

2. Thanks for mentioning this. I also get this error. `fmpy` raises the Exception when calling `fmu_instance.terminate()`. Calling `fmu_instance.freeInstance()` first gives another error when terminating. After calling the commands, however, we are still able to remove the unzipped fmu and thus properly close it. I raised an issue in `fmpy` and will update our codebase once I get a response: https://github.com/CATIA-Systems/FMPy/issues/357
