### 2022-01-06 #3861 @samanmostafavi
@FWuellhorst
I've gone through the basics so far. This includes reviewing the paper, initial installation and testing. Here's where I stand at this point:

## Comments on the paper:

**Summary:** Nicely written. I would recommend adding further references to other building energy modeling libraries (e.g., [Modelica Buildings Library](https://github.com/lbl-srg/modelica-buildings) to further emphasize the importance of developing AixCaliBuHA.

**Statement of Need:** I get the impression that ebcpy is doing the heavy lifting when it comes to creating a common interface to do the handshake between the emulator and optimization. If so, what is the statement of need specific to AixCaliBuHA? I do understand that here, you are focusing on Building Energy Models, is this the main contribution?

One more note that I would like to make is that, as you may know, there are other gradient-free model calibration methods, e.g., [EstimationPy](http://lbl-srg.github.io/EstimationPy), that are also geared towards Modelica models. I think it will be useful to mention how/if there is a plan for AixCaliBuHA to be integrated with such open source softwares (for example, it seems like you already have an API with pymoo). Having said this, I have to admit I'm not quiet sure if this is a necessary requirement in JOSS review.

Another minor issue: is there any acknowledgment of financial support necessary in this case?

## Rest of the review:

The only thing left for me is to check its functionality and more importantly, the usefulness of the code. I have done some tests and I would like to proceed by opening issues but before I do that, could you please clarify my understanding of the following:

**[Compatibility with other Libraries]:** Initially, I was very excited by the prospect of paring your software with Modelica Buildings Library. This is specifically a very relevant and useful topic since the library, alongside the ongoing [Ibpsa / project 1](https://github.com/ibpsa/project1) will provide access to an ever increasing library of building model templates. I'm not having much success here. I followed the following principle:
I substitute my fmu file but failed to load the FMU. I haven't really figured out the error yet (![error](https://user-images.githubusercontent.com/31023075/148358414-33ce0c99-063d-4f19-bea0-4cd29fbf9001.PNG)).
I looked at [examples/model](https://github.com/RWTH-EBC/AixCaliBuHA/tree/master/examples/model) and noticed I needed to do inputs/outputs specification. Since I've failed at this stage, and upon reading 'Link between optimization and class definition', I am speculating that I would need to tailor something in 'modelica_calibration_templates', Is this the case? In general, and before I open an issue, is my expectation reasonable that AixCaliBuHA is meant to be compatible with other fmus? If so, does the compatibility depends on how the fmus are compiled? I have noticed that you have claimed that usage of 'modelica_calibration_templates' is "optional".
