### 2022-01-06 #3861 @FWuellhorst
@samanmostafavi: Thanks a lot for the time invested in your initial review and your comments.

Please find my answers/comments below:

**Summary:** I agree. As the `AixLib` is part of project one, referring to all other libraries (IDEAS, Buildings, etc.) instead of only the `AixLib` is more fitting.

**Statement of Need**: The statement of need is the automated calibration. While developing the library AixCaliBuHA, we found that most interfaces can be reused for other research. For example, we used `ebcpy` for the design optimization of heat pump systems. So, `ebcpy` contains utilities and mostly abstract methods, such as the `Optimizer` and does a lot of heavy lifting regarding simulation and optimization. However, from the statement of need point of view, `AixCaliBuHA` does the "heavy" lifting of defining data classes, plotting, logging, and performing the automated calibration.

**Further gradient based packages:** I was not aware of `EstimationPy`, thank you for this recommendation. We can definitely think about integrating it into the Wrapper as well. For now, we integrate new methods If available methods fail to optimize certain tasks. As pymoo, scipy and dlib work quite quell, we have a limited priority to integrate new methods. But if you would like to see this feature, we can definitely integrate it jointly.

**Founding**: Not necessary in this case.

**Compatibility with other Libraries** It should definitely be compatible! The error you are receiving is because your FMU contains Strings, a case I encountered myself one month ago. The fix is already in the latest master. Try to install via `pip install git+https://github.com/RWTH-EBC/ebcpy`. If the error prevails, please raise an issue. A colleague was used AixCaliBuHA with the Buildings Library and EnergyPlus, so compatibility should be given.
