+++ paper.bib
@@ -108,13 +108,17 @@
     doi = {10.3384/ecp21181561}
 }
 
-@inproceedings{vering_wullhorst_ecm,
-  title={Towards an Integrated Design of Heat Pump Systems: Application and Assessment of Process Intensification using Two-Stage Optimization},
-  author={Vering, Christian and Wüllhorst, Fabian and Mehrfeld, Philipp and Müller, Dirk},
-  booktitle={{The 34th International Conference On Efficiency, Cost, Optimization, Simulation and Environmental Impact of Energy System}},
-  year={2021},
-  month = june,
-  language = {en}
+@article{vering_wullhorst_ecm,
+    title = {Towards an integrated design of heat pump systems: Application of process intensification using two-stage optimization},
+    journal = {Energy Conversion and Management},
+    volume = {250},
+    pages = {114888},
+    year = {2021},
+    issn = {0196-8904},
+    doi = {https://doi.org/10.1016/j.enconman.2021.114888},
+    url = {https://www.sciencedirect.com/science/article/pii/S0196890421010645},
+    author = {Christian Vering and Fabian Wüllhorst and Philipp Mehrfeld and Dirk Müller},
+    keywords = {Design and operation optimization, Surrogate-based optimization, Radial basis functions, Emission and cost minimization},
 }
 
 @misc{dymola_software,
@@ -131,4 +135,24 @@
   year = {2021},
   url = {https://fmi-standard.org/},
   note = {Last accessed: 24.08.2021}
+}
+
+@article{buildings,
+author = {Michael Wetter and Wangda Zuo and Thierry S. Nouidui and Xiufeng Pang},
+title = {Modelica Buildings library},
+journal = {Journal of Building Performance Simulation},
+volume = {7},
+number = {4},
+pages = {253-270},
+year  = {2014},
+publisher = {Taylor & Francis},
+doi = {10.1080/19401493.2013.765506}
+}
+
+@inproceedings{muller_aixlib_2016,
+	address = {Dresden, Germany},
+	title = {{AixLib} – {An} {Open}-{Source} {Modelica} {Library} within the {IEA}-{EBC} {Annex} 60 {Framework}},
+	booktitle = {Proceedings of the {CESBP} {Central} {European} {Symposium} on {Building} {Physics} and {BauSIM} 2016},
+	author = {Müller, D. and Lauster, M. and Constantin, A. and Fuchs, M. and Remmen, P.},
+	year = {2016}
 }
