### 2022-01-10 727d302
Update paper.bib
chore: Fix bad DOI
