+++ paper.bib
@@ -115,7 +115,7 @@
     pages = {114888},
     year = {2021},
     issn = {0196-8904},
-    doi = {https://doi.org/10.1016/j.enconman.2021.114888},
+    doi = {10.1016/j.enconman.2021.114888},
     url = {https://www.sciencedirect.com/science/article/pii/S0196890421010645},
     author = {Christian Vering and Fabian Wüllhorst and Philipp Mehrfeld and Dirk Müller},
     keywords = {Design and operation optimization, Surrogate-based optimization, Radial basis functions, Emission and cost minimization},
