### 2022-01-10 #3861 @FWuellhorst
@samanmostafavi :
I updated the references and the statement of need.
During rebuttal, your point regarding `ebcpy` became more clear to me. The old formulation "While implementing AixCaliBuHA, we identified a secondary need." is a little misleading. All features of `ebcpy` are currently only here because of the main need for the automated calibration. For a long time, `ebcpy` did not exist and `AixCaliBuHA` was the only repository. However, the classes in `ebcpy` may and already are used for other research needs. Therefore, we decided to decouple those functions from `AixCaliBuHA`. So, the research need is the automated calibration and both frameworks do their part of the "heavy lifting". We just put part of the contribution for sake of re-usability into a second framework (`ebcpy`).
I hope this statement clarifies your concerns. Thanks again for your time to review `AixCaliBuHA`.
