### 2022-01-12 RWTH-EBC/AixCaliBuHA#30 @shamsiharis
MultipleClassCalibrator Tutorial
This error only occurs for the MultipleClassCalibrator in the tutorial.ipynb
![image](https://user-images.githubusercontent.com/45209797/149229332-139e12e1-cd52-4cbc-a2f8-452e79aff2ef.png)
