### 2022-01-14 ea0ddaf
AixCaliBuHA successfully running in Parallel. Cleaning up needed  to be done.
