### 2022-01-18 f21cbd7
chore: remove semantic_release for now
