+++ fmi2.py
@@ -1,605 +1,0 @@
-""" FMI 2.0 interface """
-
-import pathlib
-from ctypes import *
-from . import free, calloc
-from .fmi1 import _FMU, FMICallException, printLogMessage
-
-
-fmi2Component            = c_void_p
-fmi2ComponentEnvironment = c_void_p
-fmi2FMUstate             = c_void_p
-fmi2ValueReference       = c_uint
-fmi2Real                 = c_double
-fmi2Integer              = c_int
-fmi2Boolean              = c_int
-fmi2Char                 = c_char
-fmi2String               = c_char_p
-fmi2Type                 = c_int
-fmi2Byte                 = c_char
-
-fmi2Status = c_int
-
-fmi2OK      = 0
-fmi2Warning = 1
-fmi2Discard = 2
-fmi2Error   = 3
-fmi2Fatal   = 4
-fmi2Pending = 5
-
-fmi2CallbackLoggerTYPE         = CFUNCTYPE(None, fmi2ComponentEnvironment, fmi2String, fmi2Status, fmi2String, fmi2String)
-fmi2CallbackAllocateMemoryTYPE = CFUNCTYPE(c_void_p, c_size_t, c_size_t)
-fmi2CallbackFreeMemoryTYPE     = CFUNCTYPE(None, c_void_p)
-fmi2StepFinishedTYPE           = CFUNCTYPE(None, fmi2ComponentEnvironment, fmi2Status)
-
-fmi2ModelExchange = 0
-fmi2CoSimulation  = 1
-
-fmi2True  = 1
-fmi2False = 0
-
-fmi2StatusKind = c_int
-
-fmi2DoStepStatus       = 0
-fmi2PendingStatus      = 1
-fmi2LastSuccessfulTime = 2
-fmi2Terminated         = 3
-
-
-def allocateMemory(nobj, size):
-    return calloc(nobj, size)
-
-
-def freeMemory(obj):
-    free(obj)
-
-
-def stepFinished(componentEnvironment, status):
-    pass
-
-
-class fmi2CallbackFunctions(Structure):
-
-    _fields_ = [('logger',               fmi2CallbackLoggerTYPE),
-                ('allocateMemory',       fmi2CallbackAllocateMemoryTYPE),
-                ('freeMemory',           fmi2CallbackFreeMemoryTYPE),
-                ('stepFinished',         fmi2StepFinishedTYPE),
-                ('componentEnvironment', fmi2ComponentEnvironment)]
-
-
-class fmi2EventInfo(Structure):
-
-    _fields_ = [('newDiscreteStatesNeeded',           fmi2Boolean),
-                ('terminateSimulation',               fmi2Boolean),
-                ('nominalsOfContinuousStatesChanged', fmi2Boolean),
-                ('valuesOfContinuousStatesChanged',   fmi2Boolean),
-                ('nextEventTimeDefined',              fmi2Boolean),
-                ('nextEventTime',                     fmi2Real)]
-
-
-class _FMU2(_FMU):
-    """ Base class for FMI 2.0 FMUs """
-
-    def __init__(self, **kwargs):
-
-        super(_FMU2, self).__init__(**kwargs)
-
-        # Inquire version numbers of header files and setting logging status
-        self._fmi2Function('fmi2GetTypesPlatform', [], [], fmi2String)
-
-        self._fmi2Function('fmi2GetVersion', [], [], fmi2String)
-
-        self._fmi2Function('fmi2SetDebugLogging',
-                           ['component', 'loggingOn', 'nCategories', 'categories'],
-                           [fmi2Component, fmi2Boolean, c_size_t, POINTER(fmi2String)])
-
-        # Creation and destruction of FMU instances and setting debug status
[…]
