### 2022-01-18 ed52946
chore: rm semantic release
