### 2022-01-18 RWTH-EBC/AixCaliBuHA#28 @FWuellhorst
To note it here: It would be possible to avoid the error at all. But, this would require manipulating the optimization, e.g.:
```
    def obj(self, xk, *args):
        self._counter += 1
        if self._counter >= self.max_itercount:
            return np.inf
```
Maybe it could also a general feature for ebcpy. For now, changing the log level will be sufficient.
