### 2022-01-18 #3861 @FWuellhorst
@fraukewiese @shamsiharis: I fixed the issue regarding `MaxIterationsReached`. To test this, please re-install via `pip install git+https://github.com/RWTH-EBC/AixCaliBuHA`.
Regarding the simulation error, I never received any logs or stats to reproduce it. Hence, I can't fix it.
