### 2022-01-21 b082e09
cleaned up, n_cpu integrated, added kwargs for pymoo (to be tested), raise ValueError for pop_size
