### 2022-01-25 #3861 @shamsiharis
@FWuellhorst Errors have been fixed. The examples all run smoothly now.

I keep getting just this small error in the **Tutorial** under **Advanced Calibration>> Multiple classes calibration**.

ERROR FMU_API: Error: Scalar system is always singular, it may be possible to evaluate parameters to avoid this, for valve.Kv = ((if valve.LinearCharacteristic then valve.kv0+(1-valve.kv0)*valve.yLim/valve.y1 else valve.kv0*exp(log(1/valve.kv0)*valve.yLim/valve.y1)))/(1.0/valve.Kv1) = 1/0

**Overall Review:** AixCaliBuHA certainly provides a functional platform to facilitate the calibration of dynamic building models. One of the highlights of the package is the use of the FMI standards to support a software-independent implementation. This can go a long way when analyzing time-series formulation of building dynamics.

The authors have included a sufficiently detailed summary highlighting the main aspects of AixCaliBuHA. I would recommend the authors to include a broad picture of the scope of AixCaliBuHA (perhaps mention this in the statement of need). Also, the summary could be improved in terms of usability over different open source simulation environments.

The tutorial runs smoothly and explains the overall process workflow. Other examples satisfactorily explain the sensitivity analysis and calibration implementations. The descriptions could be made more explicit though in terms of real-world analysis problems.
