### 2022-01-28 RWTH-EBC/AixCaliBuHA#30 @FWuellhorst
Thanks for posting the error message. I also get this error when using the latest pypi version of AixCaliBuHA. However, when installing the latest fix (https://github.com/RWTH-EBC/ebcpy/issues/45) via `pip install -e ebcpy`, the error will be gone and only a warning is printed indicating the FMU was already closed.
I forgot to mention this fix directly in the answer to your review (https://github.com/openjournals/joss-reviews/issues/3861).

If this fix does not work for you, please re-open this issue!
