### 2022-01-28 RWTH-EBC/AixCaliBuHA#32 @FWuellhorst
Update documentation of examples and paper summary
Based on the latest review (https://github.com/openjournals/joss-reviews/issues/3861), the documentation of the examples needs to be more explicit:
> The descriptions could be made more explicit though in terms of real-world analysis problems.

Also, the paper needs an update:
> I would recommend the authors to include a broad picture of the scope of AixCaliBuHA (perhaps mention this in the statement of need). Also, the summary could be improved in terms of usability over different open source simulation environments.
