### 2022-01-28 #3861 @FWuellhorst
@shamsiharis : Thanks again for the review, the comments, and the time associated with this.
The error now should be fixed, see https://github.com/RWTH-EBC/AixCaliBuHA/issues/30.
Regarding your overall review, I created an issue and will update the paper slightly to indicate that the main focus of the toolchain lies in dynamic simulations of HVAC and buildings.
One question regarding your comment:
> Also, the summary could be improved in terms of usability over different open source simulation environments.

What simulation environments do you mean? AixCaliBuHA aims at calibration, not simulation. If there is a new simulation environment that wants to use AixCaliBuHA, a `SimulationAPI` child-class could be written to enable said simulation environment. I could clarify this point in the paper if need, or did I understood your comment the wrong way?
