### 2022-01-30 de13cf8
fix (examples): Update description to include more information on the energy system behind the calibration. Also reformat examples for new example file converter.
