+++ examples/e1_A_energy_system_analysis.py
@@ -1,10 +1,10 @@
-"""
-Goals of this part of the examples:
-1. Learn how to analyze the model of your energy system
-2. Improve your `SimulationAPI` knowledge
-3. Improve your skill-set on `TimeSeriesData`
-4. Generate some measured data to later use in a calibration
-"""
+# # Example 1-A Energy system analysis
+# Goals of this part of the examples:
+# 1. Learn how to analyze the model of your energy system
+# 2. Improve your `SimulationAPI` knowledge
+# 3. Improve your skill-set on `TimeSeriesData`
+# 4. Generate some measured data to later use in a calibration
+
 # Start by importing all relevant packages
 import pathlib
 import matplotlib.pyplot as plt
@@ -19,6 +19,7 @@ def main(
 ):
     """
     Arguments of this example:
+
     :param str aixlib_mo:
         Path to the package.mo of the AixLib.
         This example was tested for AixLib version 1.0.0.
@@ -37,15 +38,18 @@ def main(
     example_path = pathlib.Path(__file__).parent
     aixcalibuha_mo = example_path.joinpath("model", "AixCaliBuHAExamples.mo")
 
-    # ######################### System analysis ##########################
-    # The best way to analyze the model which we later want to calibrate,
-    # either pause here (set a debug point) or open the models in Dymola separately.
+    # ## System analysis
+    # The best way to analyze the model which we later want to calibrate
+    # is to either pause here (set a debug point) or open the models in a GUI (OpenModelica, Dymola, o.s.).
+    # The model looks like this (only works in markdown and jupyter versions): ![img.png](./data/img_A.png)
     # Click through the system and subsystem to understand what happens in the model.
     # As you may have guessed, the analysis of an energy system can be quite complex
     # and is thus hard to automize. Before using AixCaliBuHA, you should understand
     # what happens in your system. If you have questions regarding modeling assumptions,
-    # ask e.g. the model developers of the AixLib or the IBPSA.
-    # %% Setup the Dymola-API:
+    # ask e.g. the model developers of the library you are using,
+    # in this case the AixLib.
+    #
+    # Start by setting up the Dymola-API. For more info, see the examples in ebcpy.
     dym_api = DymolaAPI(
         model_name="AixCaliBuHAExamples.HeatPumpSystemCalibration",
         cd=cd,
@@ -58,12 +62,13 @@ def main(
     )
     print("Pausing for analysis. Set the break point here if you like!")
 
-    # ######################### Data generation ##########################
+    # ## Data generation
     # We want to exemplify the process of getting experimental data using
     # the model we later want to calibrate.
     # This is a good example for two reasons:
     # 1. You really know the optimal parameters
     # 2. We don't have to deal with measurement noise etc.
+    #
     # For this example, we simulate 1 h with a 1 s sampling rate.
     # For further simulation help, check out the ebcpy examples.
     dym_api.set_sim_setup({
@@ -74,40 +79,77 @@ def main(
         return_option="savepath"
     )
 
-    # ######################### Data analysis ##########################
+    # ## Data analysis
     # Now let's analyze the data we've generated.
     # Open the file first and extract variables of interest.
-    # We want to match electrical power consumption (Pel) and room comfort (TAir)
-    # in this example.
-    # As an input of the model, TDryBulSource.y
-    # represents the outdoor air temperature
+    # As the model concerns a heat pump, we are most interested in
+    # variables relevant for heat pumps. This could be the
+    # electrical power consumption (Pel) and room comfort (TAir)
+    # supplied by the heat pump.
+    # When looking at the model, we have one input:
[…]
