+++ examples/e5_automated_process.py
@@ -1,9 +1,9 @@
-"""
-Goals of this part of the examples:
+# # Example 5 Automated process
 
-1. Learn how to run everything in one script
-"""
-
+# Goals of this part of the examples:
+# 1. Learn how to run everything in one script
+#
+# Start by importing everything
 from examples import setup_fmu, setup_calibration_classes
 from examples.e3_sensitivity_analysis_example import run_sensitivity_analysis
 from examples.e4_calibration_example import run_calibration
@@ -16,21 +16,12 @@ def main(example="A"):
     :param str example:
         Whether to use example A (requires windows) or B
     """
-
-    # Parameters for sen-analysis:
-    sim_api = setup_fmu(example=example)
-    calibration_classes, validation_class = setup_calibration_classes(
-        example=example,
-        multiple_classes=True
-    )
-
-    # Sensitivity analysis:
-    calibration_classes = run_sensitivity_analysis(sim_api=sim_api,
-                                                   cal_classes=calibration_classes)
-    # Calibration
-    run_calibration(sim_api=sim_api,
-                    cal_classes=calibration_classes,
-                    validation_class=validation_class)
+    # First we run the sensitivity analysis:
+    calibration_classes, sim_api = run_sensitivity_analysis(example=example)
+    # Then the calibration and validation
+    run_calibration(example=example,
+                    sim_api=sim_api,
+                    cal_classes=calibration_classes)
 
 
 if __name__ == '__main__':

[…]
