### 2022-01-31 229f416
chore: try without __file__
