+++ examples/docs/e2_A_optimization_problem_definition.md
@@ -1,0 +1,229 @@
+ # Example 2-A Optimization problem definition
+ Goals of this part of the examples:
+ 1. Learn how to formulate your calibration problem using our data_types
+ 2. Get to know `TunerParas`
+ 3. Get to know `Goals`
+ 4. Get to know `CalibrationClass`
+ 5. Learn how to merge multiple classes
+
+ Start by importing all relevant packages
+```python
+import pathlib
+```
+ Imports from ebcpy
+```python
+from ebcpy import TimeSeriesData
+```
+ Imports from aixcalibuha
+```python
+from aixcalibuha import TunerParas, Goals, \
+    CalibrationClass
+from aixcalibuha.data_types import merge_calibration_classes
+
+
+examples_dir
+statistical_measure = "NRMSE"
+multiple_classes = True
+```
+ ## Tuner Parameters
+ Tuner parameters are the optimization variables we will be
+ changing to match the simulated onto the measured output.
+
+ As described in the first example (e1_A_energy_system_analysis),
+ we've changed four parameters in the model. To show the usefulness
+ of sensitivity analysis prior to calibration, we will add a fifth without
+ any influence, the heating curve declination of the heat pump.
+ To define tuner parameters, you have to specify
+ - the name of the parameter
+ - an initial guess
+ - boundaries as a (min, max) tuple.
+ Note that the initial guess is not always used by optimization routines.
+ We've chosen to make it a requirement to prevent blindly accepting
+ calibration results. If the result is very far away from your initial guess
+ and you though you understand the model, maybe the parameter is just not
+ sensitive or influenced by another parameter.
+ How to load the data is up to you. To make the structure clear,
+ we use a 3 element tuple in this example:
+```python
+data = [
+    # (name, initial_value, boundaries)
+    ("heatPumpSystem.declination", 2, (1, 5)),
+    ("vol.V", 40, (1, 100)),
+    ("heaCap.C", 50000, (1000, 100000)),
+    ("rad.n", 1.24, (1, 2)),
+    ("theCon.G", 250, (1, 1000))
+]
+tuner_paras = TunerParas(
+    names=[entry[0] for entry in data],
+    initial_values=[entry[1] for entry in data],
+    bounds=[entry[2] for entry in data]
+)
+print(tuner_paras)
+print("Names of parameters", tuner_paras.get_names())
+print("Initial values", tuner_paras.get_initial_values())
+```
+ Scaling (will be done internally)
+```python
+print("Scaled initial values:\n", tuner_paras.scale(tuner_paras.get_initial_values()))
+```
+ ## Goals
+ The evaluation of your goals (or mathematically speaking 'objective function')
+ depends on the difference of measured to simulated data.
+ Thus, you need to specify both measured and simulated data.
+
+ Start by loading the measured data generated in 1_A_energy_system_analysis.py:
+```python
+data_dir = pathlib.Path(examples_dir).joinpath("data")
+meas_target_data = TimeSeriesData(data_dir.joinpath("measured_target_data.hdf"), key="example")
+```
+ Map the measured keys to the names inside your simulation
+```python
+variable_names = {
+    # Name of goal: Name of measured variable, Name of simulated variable
+    # Either use list
+    "Electricity": ["Pel", "Pel"],
+    # Or dict
+    "Room temperature": {"meas": "TAir", "sim": "vol.T"}
+}
+```
+ To match the measured data to simulated data,
+ the index has to match with the simulation output
+ Thus, convert it:
+```python
+meas_target_data.to_float_index()
+```
+ Lastly, setup the goals object. Note that the statistical_measure
+ is parameter of the python version of this example. It's a metric to
+ compare two set's of time series data. Which one to choose is up to
[…]
