+++ examples/docs/e1_A_energy_system_analysis.md
@@ -1,0 +1,162 @@
+ # Example 1-A Energy system analysis
+ Goals of this part of the examples:
+ 1. Learn how to analyze the model of your energy system
+ 2. Improve your `SimulationAPI` knowledge
+ 3. Improve your skill-set on `TimeSeriesData`
+ 4. Generate some measured data to later use in a calibration
+ Start by importing all relevant packages
+```python
+import pathlib
+import matplotlib.pyplot as plt
+```
+ Imports from ebcpy
+```python
+from ebcpy import DymolaAPI, TimeSeriesData
+
+
+examples_dir
+aixlib_mo
+cd = None
+with_plot = True
+```
+ General settings
+```python
+if cd is None:
+    cd = pathlib.Path(examples_dir).joinpath("results")
+else:
+    cd = pathlib.Path(cd)
+examples_dir = pathlib.Path(examples_dir)
+aixcalibuha_mo = examples_dir.joinpath("model", "AixCaliBuHAExamples.mo")
+```
+ ## System analysis
+ The best way to analyze the model which we later want to calibrate
+ is to either pause here (set a debug point) or open the models in a GUI (OpenModelica, Dymola, o.s.).
+ The model looks like this (only works in markdown and jupyter versions): ![img.png](../data/img_A.png)
+ Click through the system and subsystem to understand what happens in the model.
+ As you may have guessed, the analysis of an energy system can be quite complex
+ and is thus hard to automize. Before using AixCaliBuHA, you should understand
+ what happens in your system. If you have questions regarding modeling assumptions,
+ ask e.g. the model developers of the library you are using,
+ in this case the AixLib.
+
+ Start by setting up the Dymola-API. For more info, see the examples in ebcpy.
+```python
+dym_api = DymolaAPI(
+    model_name="AixCaliBuHAExamples.HeatPumpSystemCalibration",
+    cd=cd,
+    packages=[
+        aixlib_mo,
+        aixcalibuha_mo
+    ],
+    show_window=True,
+    equidistant_output=False
+)
+print("Pausing for analysis. Set the break point here if you like!")
+```
+ ## Data generation
+ We want to exemplify the process of getting experimental data using
+ the model we later want to calibrate.
+ This is a good example for two reasons:
+ 1. You really know the optimal parameters
+ 2. We don't have to deal with measurement noise etc.
+
+ For this example, we simulate 1 h with a 1 s sampling rate.
+ For further simulation help, check out the ebcpy examples.
+```python
+dym_api.set_sim_setup({
+    "stop_time": 3600,
+    "output_interval": 10
+})
+file_path = dym_api.simulate(
+    return_option="savepath"
+)
+```
+ ## Data analysis
+ Now let's analyze the data we've generated.
+ Open the file first and extract variables of interest.
+ As the model concerns a heat pump, we are most interested in
+ variables relevant for heat pumps. This could be the
+ electrical power consumption (Pel) and room comfort (TAir)
+ supplied by the heat pump.
+ When looking at the model, we have one input:
+ TDryBulSource.y, which represents the outdoor air temperature.
+ This input is important for heat pumps, as efficiency mainly depends
+ on source temperatures, and the model is of an air source heat pump.
+ We thus also extract this input for our data analysis.
+```python
+tsd = TimeSeriesData(file_path)
+tsd = tsd[["Pel", "TAir", "TDryBulSource.y"]]
+```
+ Let's check the frequency of the data to see if our 1 s sampling rate
+ was correctly applied:
+```python
+print("Simulation had index-frequency of %s with "
+      "standard deviation of %s" % tsd.frequency)
+```
[…]

[…]
