### 2022-02-01 ef92d6b
chore: Update paper based on review
