### 2022-03-04 #3861 @shamsiharis
What simulation environments do you mean? AixCaliBuHA aims at calibration, not simulation. If there is a new simulation environment that wants to use AixCaliBuHA, a SimulationAPI child-class could be written to enable said simulation environment. I could clarify this point in the paper if need, or did I understood your comment the wrong way?

@FWuellhorst I apologize I completely missed your comment.

Yeah! You understood it correctly. It is not a major comment. You can clarify this in the paper.

I definitely look forward to using the package in my future research.

@fraukewiese Most welcome. I really enjoyed reviewing the package.
