### 2022-03-07 #3861 @samanmostafavi
@FWuellhorst
I have tested the software. Here's my review:

## Software documentation:
I think [this page](https://ebc.pages.rwth-aachen.de/EBC_all/github_ci/AixCaliBuHA/master/docs/index.html) can be significantly improved by adding more images (results, graph demonstrating workflow, etc.) and explaining the methods of choice, even if the majority of it will be done by citing other works. In addition, the dataframe, relationship to ebpcy, all the different knobs for SA and calibration etc. needs to be explained. I believe most of this already exists in the Jupyter Notebook example and the accompanying paper.

## Examples:
KUDOS! I really like how clean (and fool-proof) this part is. A more detailed analysis is provided in comments on the tutorial ( which is basically the same example in a notebook format).
I only have a single complaint here: I think there needs to be a "useful" example for a relatively big model that would showcase the relevance of SA+Calibration workflow. For example you can show how calibration will improve. (my two cents would be to pick something from Modelica Buildings Library/IDEAS).

## Comments on the tutorial:

**Summary:**
Nice step by step tutorial. ***Basics*** are adequately laid out. The existing workflow in AixCaliBuHA is not intuitive to the untrained eye. Having said that, once the user is patient enough to follow the instructions, I found no flaws whatsoever with the workflow. Full scores for that!

***On Sensitivity Analysis*** Sobol sensitivity analysis will not scale for models with large number of parameters (too many samples and model evaluation are required). I also suspect this might be the case for Morris method. The authors should at least acknowledge this. Any suggestion on how to deal with such cases? Any prospect for parallelization for example?
Finally, GSA methods has limitation, specially when there is inter-correlations between model parameters which will likely be the case in complex models. In these scenarios, prior assumption about models might greatly assist with better SA.  Unless I am missing something here, AixCaliBuHA does not have this capability. Please confirm if otherwise.

***On Calibration***  Very nice visualization! I generally found the plots well done and very useful.
I am not convinced with black box optimization but I do understand the limitations of calibrating an FMU (which is effectively a black-box). It is likely that gradient-free methods will by and large not work well for large complex model.

**Future work suggestion**: While not necessary the scope of this software, I strongly encourage the authors to consider future workflows that include parsing equations out of Modelica models (xml files) (e.g., [pymoca](https://github.com/pymoca/pymoca)) that enables use of automatic differentiation for gradient-based optimization.

## Final thought (and concern):
The software can potentially be a useful package for certain black-box optimization tasks. The tool will specially be very useful for analyzing and calibrating a growing family of small(ish) Modelica buildings models.
I reserve my judgment on whether or not this will also be the case for over parameterized complex buildings models.
