### 2022-03-18 #3861 @FWuellhorst
@samanmostafavi and @shamsiharis: First of all, thank you both for your review, your comments, and the time associated with this!

@shamsiharis: Thanks, I will clarify this in the paper.

@samanmostafavi:

## Documentation
I agree, I updated the page (and thus also the README.md).

## Paper
Commercial software for calibration (mainly provided inside Dymola) is not highlighted, as they are not open-source.
I added existing options for calibration of FMUs and explained how AixCaliBuHA may integrate them.

## Examples
Thanks for the kudos.
Regarding the detailed example: We are currently in the process of performing a complex calibration of a hydraulic system based on IBPSA models. This will highlight the usefulness of the sensitivity analysis. Afterward, we will add this complex example to the existing examples as well.
The current example at least shows how less sensitive parameters are eliminated and thus reduce the optimization effort. If required, we can extend this by adding further tuner parameters prior to SA.

Furthermore, we encourage any users to add their use-cases to the list of examples.

## Tutorial

### Summary
It's good to hear that the workflow is clear through the tutorials, thanks for the comment.

### Sensitivity Analysis

Regarding scaling: Parallelization will be added in future versions for the GSA. As our SimulationAPI is already enabled for parallelization, this is a quick feature.
Regarding limits of GSA: AixCaliBuHA does not aim to improve GSA methods but rather uses them. The same goes for the calibration (see below). I will highlight this in the paper.
Regarding suggestions for better model assumptions, do you know a software that is capable of doing this? We are not aware of existing solutions. If there is one, we will happily couple it with AixCaliBuHA.

### Calibration:
Thanks!

Regarding black-box optimization:
Coupled HVAC and building systems are quite complex, including nonlinear controls with state-events as well as stiff equations leading to MINLPs.
For these systems, we had limited success using export to enable gradient-based methods using the Optimica compiler. We will try out `pymoca` in future studies as well.
However, even after export, gradient-based methods for global optimization of MINLP require substantial knowledge in numerical optimization.
In contrast, black-box methods may not get the global optimum, but achieve promising results with little knowledge of the system.
I fully agree that researchers should first try export and gradient-based methods.
However, if this fails, AixCaliBuHA will help to at least improve the calibration quality.

Regarding the **Future work suggestion**:
If `pymoca` works for example models, we will include an option to automatically export the models to `pymoca` and link them to gradient-based calibration toolchains.

## Final thought (and concern)

The usefulness in over-parameterized complex buildings models depends mostly on the optimization algorithm. As the number of options is easy to extend, AixCaliBuHA can adapt to possibly challenging calibration tasks.
We used the toolchain for complex models (HighOrder Building and detailed HVAC components) already with success.
If two tuner parameters exist that influence the outputs in opposite directions (e.g. internal gains and wall insulation), the expert definitely has to set useful boundaries for these optimization variables.
As this is in line with your concern regarding GSA and calibration, I will empathize in the paper, that AixCaliBuHA does not substitute expert knowledge. It helps experts by automating tedious and time-consuming tasks.

@fraukewiese, @samanmostafavi, @shamsiharis : I added all requested changes into the existing PR https://github.com/RWTH-EBC/AixCaliBuHA/pull/33. If the changes are accepted, I will happily merge this PR.
