### 2022-03-28 #3861 @samanmostafavi
@fraukewiese I think my concerns are addressed. Both package and documentation are in a good state. My only remaining concern is that the toolchain might not be as useful for bigger problems. Having said that, and as elaborated in the response above by @FWuellhorst, the case by case, or rather a _blanket_, usefulness is out of the scope of this software and the developing team have done a reasonably good job with good software engineering practices that I think, and hope, will come to the aid of modelers in the near future.

@FWuellhorst It has been a pleasure reviewing your tool. I wish you the best and hope you keep up the good work in developing tools that make building modelers lives easier!  I will definitely be following :-)

Cheers,
Saman

P.S. @FWuellhorst yes, parsing tools such as pymoca will likely have very little chance of being applicable to complex hybrid models with thousands of equations. Having played around with it over the past few weeks, I think it's not worth the trouble for this domain of problems.
