### 2022-03-28 #3861 @FWuellhorst
@samanmostafavi Thanks for the really nice review! I will definitely try out the parsing tools. Especially for new approaches such as agent based control or digital twins, small models will be more relevant for us. For these cases, pymoca will hopefully work! I will also follow your updates :)
