### 2022-04-04 f10e2f4
feat: add parallelization for SA as well #20
