### 2022-04-04 bed9f89
chore: Minor adjustments to use n_cpus everytime it is given.
