+++ aixcalibuha/calibration/calibrator.py
@@ -196,7 +196,7 @@ class Calibrator(Optimizer):
         :rtype: float
         """
         # Info: This function is called by the optimization framework (scipy, dlib, etc.)
-        #%% Initialize class objects
+        # Initialize class objects
         self._current_iterate = xk
         self._counter += 1
 
@@ -240,17 +240,96 @@ class Calibrator(Optimizer):
             )
             return self.ret_val_on_error
 
-        self.goals.set_sim_target_data(sim_target_data)
+        total_res = self._kpi_and_logging_calculation(
+            xk_descaled=xk_descaled,
+            counter=self._counter,
+            results=sim_target_data
+        )
+
+        return total_res
+
+    def mp_obj(self, x, *args):
+        # Initialize list for results
+        num_evals = len(x)
+        total_res_list = np.empty([num_evals, 1])
+        # Set initial values of variable and fixed parameters
+        self.sim_api.result_names = self.goals.get_sim_var_names()
+        initial_names = self.tuner_paras.get_names()
+        parameters = self.fixed_parameters.copy()
+
+        parameter_list = []
+        xk_descaled_list = []
+        for _xk_single in x:
+            # Convert set if multiple goals of different scales are used
+            xk_descaled = self.tuner_paras.descale(_xk_single)
+            xk_descaled_list.append(xk_descaled)
+            # Update Parameters
+            parameter_copy = parameters.copy()
+            parameter_copy.update({name: value for name, value in zip(initial_names, xk_descaled.values)})
+            parameter_list.append(parameter_copy)
+
+        # Simulate
+        self.logger.log(f"Starting {num_evals} simulations on {self.sim_api.n_cpu} cores", level=logging.INFO)
+        if self.save_files:
+            result_file_names = [f"simulation_{self._counter + idx}" for idx in range(len(parameter_list))]
+            _filepaths = self.sim_api.simulate(
+                parameters=parameter_list,
+                return_option="savepath",
+                savepath=self.sim_api.cd,
+                result_file_name=result_file_names,
+                fail_on_error=self.fail_on_error,
+                inputs=self.calibration_class.inputs,
+                **self.calibration_class.input_kwargs
+            )
+            # Load results
+            results = []
+            for _filepath in _filepaths:
+                if _filepath is None:
+                    results.append(None)
+                else:
+                    results.append(data_types.TimeSeriesData(_filepath))
+        else:
+            results = self.sim_api.simulate(
+                parameters=parameter_list,
+                inputs=self.calibration_class.inputs,
+                fail_on_error=self.fail_on_error,
+                **self.calibration_class.input_kwargs
+            )
+        self.logger.log(f"Finished {num_evals} simulations", level=logging.INFO)
+        for idx, result in enumerate(results):
+            self._counter += 1
+            self._current_iterate = result
+            if result is None:
+                total_res_list[idx] = self.ret_val_on_error
+                continue
+            total_res = self._kpi_and_logging_calculation(
+                xk_descaled=xk_descaled_list[idx],
+                counter=self._counter,
+                results=result
+            )
+            # Add single objective to objective list of total Population
+            total_res_list[idx] = total_res
+
+        return total_res_list
+
+    def _kpi_and_logging_calculation(self, *, xk_descaled, counter, results):
+        """
+        Function to calculate everything needed in the obj or mp_obj
+        function after the simulation finished.
+
+        """
+        xk = self.tuner_paras.scale(xk_descaled)
+
+        self.goals.set_sim_target_data(results)
         # Trim results based on start and end-time of cal class
         self.goals.set_relevant_time_intervals(self.calibration_class.relevant_intervals)
 
-        #%% Evaluate the current objective
+        # %% Evaluate the current objective
[…]
