### 2022-04-04 e18c51b
chore: increase version number. Due to breaking change by 0.1
