### 2022-04-06 d1d88c1
chore: move logging to ebcpy
