### 2022-04-07 dd70e66
chore: adjusted default pop-size for pymoo ga
