### 2022-04-11 83868bc
chore: fix bad examples_dir parameter and uniform args
