### 2022-04-20 #3861 @fraukewiese
@FWuellhorst : Please check the following in the paper:

- Is there any DOI or URL available for King, D. E. (2009). Dlib-ml: A machine learning toolkit. Journal of Machine Learning
Research, 10, 1755–1758 ?
- Is there any DOI or URL available for  Mehrfeld, P., Nürenberg, M., & Müller, D. (2021). Model Calibration of an Air Source Heat Pump System for Transient Simulations in Modelica. 13th IEA Heat Pump Conference ?
- Is there any DOI or URL available for Müller, D., Lauster, M., Constantin, A., Fuchs, M., & Remmen, P. (2016). AixLib – An Open-Source Modelica Library within the IEA-EBC Annex 60 Framework. Proceedings of the CESBP Central European Symposium on Building Physics and BauSIM 2016 ?
- Same here: Vering, C., Borges, D., Sebastian Coakly, Krützfeldt, H., Mehrfeld, P., & Müller, D. (2021). Digital twin design with on-line calibration for HVAC systems in buildings. Proceedings of the 17th IBPSA Conference, 8.
