+++ paper.bib
@@ -41,7 +41,6 @@
   journal = {Journal of Machine Learning Research},
   year = {2009},
   volume = {10},
-  doi = {10.1145/1577069.1755843},
   pages = {1755-1758},
 }
 
