### 2022-04-20 #3861 @FWuellhorst
@fraukewiese
Sadly, these conference papers don't have a DOI. Regarding "King, D. E. (2009). Dlib-ml: A machine learning toolkit. Journal of Machine Learning Research, 10, 1755–1758", a DOI is available on [ReserachGate](https://www.researchgate.net/publication/220320646_Dlib-ml_A_Machine_Learning_Toolkit) (10.1145/1577069.1755843), however, this DOI is not valid according to https://dx.doi.org/ and thus also for the editorial bot.

Is this a problem? If not, I would continue with the described steps.
