### 2022-04-20 053c43a
fix: Validation bug. Rm unwanted tests as well (not a valid unittest)
