+++ examples/e4_calibration_example.py
@@ -147,7 +147,7 @@ def run_calibration(
     # Start the validation process
     modelica_calibrator.validate(
         validation_class=validation_class,
-        tuner_parameter_values=list(result.values())
+        calibration_result=result
     )
     # Don't forget to close the simulation api:
     sim_api.close()
