### 2022-04-21 #3861 @fraukewiese
@FWuellhorst
I see. Then it is ok to go on without those doi.
