### 2022-04-21 #3861 @FWuellhorst
@fraukewiese : Thanks for the quick answer and the time associated with this review process!
I followed all steps:
- Tagged release: https://github.com/RWTH-EBC/AixCaliBuHA/releases/tag/0.3.0
- Release tag: v_0.3.0
- DOI of zenodo: 10.5281/zenodo.6475439

I hope that everything is correct.
