### 2022-04-23 #3861 @FWuellhorst
@samanmostafavi @shamsiharis Thanks again for the really good review and all your time put into this!

@fraukewiese @arfon: Thanks for providing the possibility to publish open-source software for research using JOSS. I really appreciate the concept of the journal and the review process. I've signed up as a reviewer. Maybe I can be of help in future reviews as well.
